requirements.txt
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26
pydantic>=2.7
pydantic-settings>=2.3
diskcache>=5.6
tenacity>=8.5
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-exporter-otlp>=1.25
opentelemetry-exporter-console>=1.25
typer>=0.12
rich>=13.7
pypdf>=4.2

# Service / deployment
fastapi>=0.112
uvicorn[standard]>=0.30

# Async + rate limiting
aiolimiter>=1.2

# Token estimation
tiktoken>=0.7

# Testing / eval
pytest>=8.2
pytest-asyncio>=0.23
